Character Creation was simplified slightly, as neccessitated by a number of boxes not being information this project is dealing with (cosmetic descriptions, and equipment being a tabletop-purchased thing)
Encounter diagram was simplifed based on the conditional "unaware actors do not get actions." and as such eliminated a branch.

usecase diagrams were nudged a bit and should not have any real changes.

Character generation sequence diagram was not touched yet (other than a few typos) due to a failing of caffine.